pkg/payload/precondition/clusterversion: New file for RecentEtcdBackup
There's enough going on in the RecentEtcdBackup and Upgradeable
preconditions that having them in their own files doesn't feel too
empty.
